+++ src/main/java/b22/metro2033/Entity/Army/Post.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
     private String location;
 
@@ -23,17 +23,17 @@ public class Post {
     public Post() {
     }
 
-    public Post(int id, String name, String location) {
+    public Post(long id, String name, String location) {
         this.id = id;
         this.name = name;
         this.location = location;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
